Smaller data structures reflecting limits of OB3/JM formats

+++ src/world.c
@@ -1053,7 +1053,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                 if (colour != colour_1 || i17 != 0) {
                     if (direction == 0) {
                         if (colour != COLOUR_TRANSPARENT) {
-                            int *colour_vertices = calloc(3, sizeof(int));
+                            uint16_t *colour_vertices = calloc(3, sizeof(uint16_t));
 
                             colour_vertices[0] = r_y + r_x * 96 + 96;
                             colour_vertices[1] = r_y + r_x * 96;
@@ -1071,7 +1071,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                         }
 
                         if (colour_1 != COLOUR_TRANSPARENT) {
-                            int *colour_1_vertices = calloc(3, sizeof(int));
+                            uint16_t *colour_1_vertices = calloc(3, sizeof(uint16_t));
 
                             colour_1_vertices[0] = r_y + r_x * 96 + 1;
                             colour_1_vertices[1] = r_y + r_x * 96 + 96 + 1;
@@ -1089,7 +1089,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                         }
                     } else {
                         if (colour != COLOUR_TRANSPARENT) {
-                            int *colour_vertices = calloc(3, sizeof(int));
+                            uint16_t *colour_vertices = calloc(3, sizeof(uint16_t));
 
                             colour_vertices[0] = r_y + r_x * 96 + 1;
                             colour_vertices[1] = r_y + r_x * 96 + 96 + 1;
@@ -1107,7 +1107,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                         }
 
                         if (colour_1 != COLOUR_TRANSPARENT) {
-                            int *colour_1_vertices = calloc(3, sizeof(int));
+                            uint16_t *colour_1_vertices = calloc(3, sizeof(uint16_t));
 
                             colour_1_vertices[0] = r_y + r_x * 96 + 96;
                             colour_1_vertices[1] = r_y + r_x * 96;
@@ -1125,7 +1125,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                         }
                     }
                 } else if (colour != COLOUR_TRANSPARENT) {
-                    int *vertices = calloc(4, sizeof(int));
+                    uint16_t *vertices = calloc(4, sizeof(int));
 
                     vertices[0] = r_y + r_x * 96 + 96;
                     vertices[1] = r_y + r_x * 96;
@@ -1153,7 +1153,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                     game_data_tile_type[decoration - 1] == BRIDGE_TILE_TYPE) {
                     int fill_front = game_data_tile_decoration[decoration - 1];
 
-                    int *vertices = calloc(4, sizeof(int));
+                    uint16_t *vertices = calloc(4, sizeof(uint16_t));
 
                     vertices[0] = game_model_vertex_at(
                         game_model, r_x * TILE_SIZE,
@@ -1198,7 +1198,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                         int fill_front =
                             game_data_tile_decoration[decoration_south - 1];
 
-                        int *vertices = calloc(4, sizeof(int));
+                        uint16_t *vertices = calloc(4, sizeof(uint16_t));
 
                         vertices[0] = game_model_vertex_at(
                             game_model, r_x * TILE_SIZE,
@@ -1245,7 +1245,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                                                           world, r_x, r_y - 1) -
                                                       1];
 
-                        int *vertices = calloc(4, sizeof(int));
+                        uint16_t *vertices = calloc(4, sizeof(uint16_t));
 
                         vertices[0] = game_model_vertex_at(
                             game_model, r_x * TILE_SIZE,
@@ -1290,7 +1290,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                         int fill_front =
                             game_data_tile_decoration[decoration_east - 1];
 
-                        int *vertices = calloc(4, sizeof(int));
+                        uint16_t *vertices = calloc(4, sizeof(uint16_t));
 
                         vertices[0] = game_model_vertex_at(
                             game_model, r_x * TILE_SIZE,
@@ -1335,7 +1335,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                         int face_fill =
                             game_data_tile_decoration[decoration_west - 1];
 
-                        int *vertices = calloc(4, sizeof(int));
+                        uint16_t *vertices = calloc(4, sizeof(uint16_t));
 
                         vertices[0] = game_model_vertex_at(
                             game_model, r_x * TILE_SIZE,
@@ -1788,7 +1788,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
 
             if (diagonal > 12000 && diagonal < 24000 &&
                 world_get_wall_roof(world, r_x - 1, r_y - 1) == 0) {
-                int *vertices = calloc(3, sizeof(int));
+                uint16_t *vertices = calloc(3, sizeof(uint16_t));
 
                 vertices[0] =
                     game_model_vertex_at(world->parent_model, vertex_4_x,
@@ -1806,7 +1806,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                                        roof_id, COLOUR_TRANSPARENT);
             } else if (diagonal > 12000 && diagonal < 24000 &&
                        world_get_wall_roof(world, r_x + 1, r_y + 1) == 0) {
-                int *vertices = calloc(3, sizeof(int));
+                uint16_t *vertices = calloc(3, sizeof(uint16_t));
 
                 vertices[0] =
                     game_model_vertex_at(world->parent_model, vertex_1_x,
@@ -1825,7 +1825,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
             } else if (world_get_wall_diagonal(world, r_x, r_y) > 0 &&
                        world_get_wall_diagonal(world, r_x, r_y) < 12000 &&
                        world_get_wall_roof(world, r_x + 1, r_y - 1) == 0) {
-                int *vertices = calloc(3, sizeof(int));
+                uint16_t *vertices = calloc(3, sizeof(uint16_t));
 
                 vertices[0] =
                     game_model_vertex_at(world->parent_model, vertex_3_x,
@@ -1844,7 +1844,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
             } else if (world_get_wall_diagonal(world, r_x, r_y) > 0 &&
                        world_get_wall_diagonal(world, r_x, r_y) < 12000 &&
                        world_get_wall_roof(world, r_x - 1, r_y + 1) == 0) {
-                int *vertices = calloc(3, sizeof(int));
+                uint16_t *vertices = calloc(3, sizeof(uint16_t));
 
                 vertices[0] =
                     game_model_vertex_at(world->parent_model, vertex_2_x,
@@ -1862,7 +1862,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                                        roof_id, COLOUR_TRANSPARENT);
             } else if (terrain_height == terrain_east_height &&
                        terrain_south_east_height == terrain_south_height) {
-                int *vertices = calloc(4, sizeof(int));
+                uint16_t *vertices = calloc(4, sizeof(uint16_t));
 
                 vertices[0] =
                     game_model_vertex_at(world->parent_model, vertex_1_x,
@@ -1884,7 +1884,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                                        roof_id, COLOUR_TRANSPARENT);
             } else if (terrain_height == terrain_south_height &&
                        terrain_east_height == terrain_south_east_height) {
-                int *vertices = calloc(4, sizeof(int));
+                uint16_t *vertices = calloc(4, sizeof(uint16_t));
 
                 vertices[0] =
                     game_model_vertex_at(world->parent_model, vertex_3_x,
@@ -1918,7 +1918,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                 }
 
                 if (!direction) {
-                    int *triangle_1 = calloc(3, sizeof(int));
+                    uint16_t *triangle_1 = calloc(3, sizeof(uint16_t));
 
                     triangle_1[0] =
                         game_model_vertex_at(world->parent_model, vertex_2_x,
@@ -1935,7 +1935,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                     game_model_create_face(world->parent_model, 3, triangle_1,
                                            roof_id, COLOUR_TRANSPARENT);
 
-                    int *triangle_2 = calloc(3, sizeof(int));
+                    uint16_t *triangle_2 = calloc(3, sizeof(uint16_t));
 
                     triangle_2[0] =
                         game_model_vertex_at(world->parent_model, vertex_3_x,
@@ -1952,7 +1952,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                     game_model_create_face(world->parent_model, 3, triangle_2,
                                            roof_id, COLOUR_TRANSPARENT);
                 } else {
-                    int *triangle_1 = calloc(3, sizeof(int));
+                    uint16_t *triangle_1 = calloc(3, sizeof(uint16_t));
 
                     triangle_1[0] =
                         game_model_vertex_at(world->parent_model, vertex_1_x,
@@ -1967,7 +1967,7 @@ void world_load_section_from4(World *world, int x, int y, int plane,
                     game_model_create_face(world->parent_model, 3, triangle_1,
                                            roof_id, COLOUR_TRANSPARENT);
 
-                    int *triangle_2 = calloc(3, sizeof(int));
+                    uint16_t *triangle_2 = calloc(3, sizeof(uint16_t));
 
                     triangle_2[0] = game_model_vertex_at(
                         world->parent_model, vertex_4_x,
@@ -2135,7 +2135,7 @@ void world_create_wall(World *world, GameModel *game_model, int wall_object_id,
         vertex_y2 -= delta_y;
     }
 
-    int *vertices = calloc(4, sizeof(int));
+    uint16_t *vertices = calloc(4, sizeof(uint16_t));
 
     vertices[0] = game_model_vertex_at(
         game_model, vertex_x1, -world->terrain_height_local[x1][y1], vertex_y1);
@@ -2162,7 +2162,7 @@ void world_create_wall(World *world, GameModel *game_model, int wall_object_id,
     }
 
     if (thick_walls) {
-        int *parallel_vertices = calloc(4, sizeof(int));
+        uint16_t *parallel_vertices = calloc(4, sizeof(uint16_t));
 
         int parallel_vertex_x1 = x1 * TILE_SIZE;
         int parallel_vertex_y1 = y1 * TILE_SIZE;
@@ -2203,7 +2203,7 @@ void world_create_wall(World *world, GameModel *game_model, int wall_object_id,
             game_model->face_tag[parallel_wall_face] = 0;
         }
 
-        int *top_vertices = calloc(4, sizeof(int));
+        uint16_t *top_vertices = calloc(4, sizeof(uint16_t));
         top_vertices[0] = vertices[1];
         top_vertices[1] = parallel_vertices[1];
         top_vertices[2] = parallel_vertices[2];
